refactor: move add_days and add_points for membership module
